* Fixed unhandled exception when unchecking an SFX panel box when the SFX section is not present.

+++ TranslationScriptMaker/TranslationScriptMaker/RawsViewerForm.cs
@@ -804,7 +804,10 @@ namespace TranslationScriptMaker
 					++linesBetweenSFX;
 				}
 
-				pageContents.RemoveRange(indexToRemove, linesBetweenSFX);
+				if ( indexToRemove >= 0 )
+				{
+					pageContents.RemoveRange(indexToRemove, linesBetweenSFX);
+				}
 			}
 
 			UpdateScriptEditorContents(pageContents);
